# refactor: remove atomic op in #591