--- a/foyer-memory/src/prelude.rs
+++ b/foyer-memory/src/prelude.rs
@@ -13,9 +13,9 @@
 //  limitations under the License.
 
 pub use crate::{
-    cache::{Cache, CacheBuilder, CacheEntry, EvictionConfig, Fetch, FetchState},
+    cache::{Cache, CacheBuilder, CacheEntry, EvictionConfig, Fetch},
     context::CacheContext,
     eviction::{fifo::FifoConfig, lfu::LfuConfig, lru::LruConfig, s3fifo::S3FifoConfig},
-    generic::Weighter,
+    generic::{FetchState, Weighter},
 };
 pub use ahash::RandomState;