--- a/foyer/src/hybrid/cache.rs
+++ b/foyer/src/hybrid/cache.rs
@@ -35,6 +35,7 @@ use foyer_common::{
 };
 use foyer_memory::{Cache, CacheContext, CacheEntry, Fetch, FetchState};
 use foyer_storage::{DeviceStats, Storage, Store};
+use futures::FutureExt;
 use minitrace::prelude::*;
 use pin_project::pin_project;
 use tokio::sync::oneshot;
@@ -402,9 +403,8 @@ where
     S: HashBuilder + Debug,
 {
     #[pin]
-    inner: Fetch<K, V, anyhow::Error, S>,
+    inner: Fetch<K, V, anyhow::Error, S, bool>,
 
-    enqueue: Arc<AtomicBool>,
     storage: Store<K, V, S>,
 }
 
@@ -417,11 +417,11 @@ where
     type Output = anyhow::Result<CacheEntry<K, V, S>>;
 
     fn poll(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Self::Output> {
-        let this = self.project();
-        let res = ready!(this.inner.poll(cx));
+        let mut this = self.project();
+        let res = ready!(this.inner.as_mut().poll(cx));
 
         if let Ok(entry) = res.as_ref() {
-            if this.enqueue.load(Ordering::Acquire) {
+            if *this.inner.ext() {
                 this.storage.enqueue(entry.clone(), false);
             }
         }
@@ -436,7 +436,7 @@ where
     V: StorageValue,
     S: HashBuilder + Debug,
 {
-    type Target = Fetch<K, V, anyhow::Error, S>;
+    type Target = Fetch<K, V, anyhow::Error, S, bool>;
 
     fn deref(&self) -> &Self::Target {
         &self.inner
@@ -477,14 +477,14 @@ where
         let now = Instant::now();
 
         let store = self.storage.clone();
-        let enqueue = Arc::<AtomicBool>::default();
+
         let future = fetch();
         let inner = self.memory.fetch_inner(
             key.clone(),
             context,
             || {
                 let metrics = self.metrics.clone();
-                let enqueue = enqueue.clone();
+
                 async move {
                     match store.load(&key).await.map_err(anyhow::Error::from)? {
                         None => {}
@@ -493,16 +493,15 @@ where
                             metrics.hybrid_hit.increment(1);
                             metrics.hybrid_hit_duration.record(now.elapsed());
 
-                            return Ok(v);
+                            return Ok((v, false));
                         }
                     }
 
                     metrics.hybrid_miss.increment(1);
                     metrics.hybrid_miss_duration.record(now.elapsed());
 
-                    enqueue.store(true, Ordering::Release);
-
                     future
+                        .map(|res| res.map(|v| (v, true)))
                         .in_span(Span::enter_with_local_parent("foyer::hybrid::fetch::fn"))
                         .await
                         .map_err(anyhow::Error::from)
@@ -518,7 +517,6 @@ where
 
         let inner = HybridFetchInner {
             inner,
-            enqueue,
             storage: self.storage.clone(),
         };