@MrCroxx 2024-07-01
# Benchmark

```bash
rm -rf /6536/foyer && cargo build --release && RUST_BACKTRACE=1 RUST_LOG=info ./target/release/foyer-bench --dir /6536/foyer --mem 64 --disk 102400 --file-size 64 --get-range 10000 --flushers 4 --reclaimers 4 --time 10 --writers 256 --w-rate 4 --admission-rate-limit 500 --readers 1024 --r-rate 1 --runtime --compression none --distribution zipf --distribution-zipf-s 0.8 --metrics
```

## main

```
Total:
disk total iops: 3477.8
disk total throughput: 583.5 MiB/s
disk read iops: 1245.0
disk read throughput: 82.7 MiB/s
disk write iops: 2232.8
disk write throughput: 500.8 MiB/s
insert iops: 14278.7/s
insert throughput: 892.5 MiB/s
insert lat p50: 4us
insert lat p90: 8us
insert lat p99: 19us
insert lat p999: 48us
insert lat p9999: 176us
insert lat p99999: 349us
insert lat pmax: 739us
get iops: 17325.1/s
get miss: 8.56% 
get throughput: 1.0 GiB/s
get hit lat p50: 3us
get hit lat p90: 919us
get hit lat p99: 2975us
get hit lat p999: 4735us
get hit lat p9999: 5919us
get hit lat p99999: 6943us
get hit lat pmax: 8511us
get miss lat p50: 46us
get miss lat p90: 1103us
get miss lat p99: 2943us
get miss lat p999: 4255us
get miss lat p9999: 5311us
get miss lat p99999: 6015us
get miss lat pmax: 6015us
```

## PR

```
Total:
disk total iops: 3477.8
disk total throughput: 583.5 MiB/s
disk read iops: 1245.0
disk read throughput: 82.7 MiB/s
disk write iops: 2232.8
disk write throughput: 500.8 MiB/s
insert iops: 14278.7/s
insert throughput: 892.5 MiB/s
insert lat p50: 4us
insert lat p90: 8us
insert lat p99: 19us
insert lat p999: 48us
insert lat p9999: 176us
insert lat p99999: 349us
insert lat pmax: 739us
get iops: 17325.1/s
get miss: 8.56% 
get throughput: 1.0 GiB/s
get hit lat p50: 3us
get hit lat p90: 919us
get hit lat p99: 2975us
get hit lat p999: 4735us
get hit lat p9999: 5919us
get hit lat p99999: 6943us
get hit lat pmax: 8511us
get miss lat p50: 46us
get miss lat p90: 1103us
get miss lat p99: 2943us
get miss lat p999: 4255us
get miss lat p9999: 5311us
get miss lat p99999: 6015us
get miss lat pmax: 6015us
```